+++ src/lib.rs
@@ -5,7 +5,6 @@ use std::{
     path::{Path, PathBuf},
     str,
 };
-pub mod base16;
 pub mod commands;
 pub mod diffs;
 pub mod errors;
